% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{DataMa}
\alias{DataMa}
\alias{DataMa-package}
\title{DataMa: magic functions for data management}
\description{
A coolection of data management tools
}
\details{
A collection of functions which might be helpful/useful for our
day-to-day ventures into data analysis at Environmentalinformatics
group, Department of Geography, Philipps University Marburg.
}
\author{
Thomas Nauss
\emph{Maintainer:} Thomas Nauss \email{nausst@gmail.com}
}
\keyword{package}

