% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiply}
\alias{mult}
\alias{multiply}
\title{Multiplication of two data sets if they are numeric}
\usage{
multiply(ds1, ds2)
}
\arguments{
\item{ds1}{data set 1}

\item{ds2}{data set 2}
}
\value{
A vector containing the product of data set 1 and 2.
}
\description{
Multiplication of two data sets if they are numeric
}
\details{
The two data sets are multiplied if they are numeric.
}
\note{
The data sets must be of type vector.
}
\examples{
# Multiply two data sets

a <- c(1,2,3,4,5)
b <- c(10,20,30,40,50)
example <- multiply(a, b)
print(example)
}
\author{
Thomas Nauss
}
\seealso{
For further functions something else.
}

